CMakeLists: add src/ to build interface include directories
The src directory contains some Fortran and C include files, which are
added to the PUBLIC_HEADER property of the tdpack-static target via
PROJECT_INCLUDE_FILES. They are then installed along with the package
since the install command installs PUBLIC_HEADER to 'include/'.

They are however _not_ included in the INTERACE_INCLUDE_DIRECTORIES
property, such that when using 'add_subdirectory' to build tdpack as a
CMake subdirectory (as is done in GEM under WITH_SYSTEM_RPN=FALSE),
these headers are not found. Adjust INTERFACE_INCLUDE_DIRECTORIES to add
their location, making use of the PROJECT_SOURCE_DIR variable [1] which
makes sure that we always point to the root of the tdpack sources (using
CMAKE_SOURCE_DIR would not work since in the GEM build it refers to the
root of the GEM sources).

[1] https://cmake.org/cmake/help/latest/variable/PROJECT_SOURCE_DIR.html

Helped-by: Philippe Carphin <philippe.carphin2@canada.ca>